requirements.txt
pytest
flake8
attrs
autobahn
twisted
